Misc castings etc to deal with WIndows warnings

+++ trunk/src/gmt_io.c
@@ -1,5 +1,5 @@
 /*--------------------------------------------------------------------
- *	$Id: gmt_io.c,v 1.156 2008-03-27 09:35:30 guru Exp $
+ *	$Id: gmt_io.c,v 1.157 2008-04-25 04:00:40 guru Exp $
  *
  *	Copyright (c) 1991-2008 by P. Wessel and W. H. F. Smith
  *	See COPYING file for copying and redistribution conditions.
@@ -2939,7 +2939,7 @@ int GMT_import_table (void *source, int source_type, struct GMT_TABLE **table, d
 					T->segment[seg]->coord[GMT_Y][row] = T->segment[seg]->coord[GMT_Y][0];
 					T->segment[seg]->n_rows++;
 				}
-				for (row = 0; row < (size_t)T->segment[seg]->n_rows - 1; row++) {
+				for (row = 0; row < T->segment[seg]->n_rows - 1; row++) {
 					dlon = T->segment[seg]->coord[GMT_X][row+1] - T->segment[seg]->coord[GMT_X][row];
 					if (fabs (dlon) > 180.0) dlon = copysign (360.0 - fabs (dlon), -dlon);	/* Crossed Greenwhich or Dateline, pick the shortest distance */
 					lon_sum += dlon;
